requirements.txt
appnope==0.1.4
asttokens==3.0.1
certifi==2025.11.12
charset-normalizer==3.4.4
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
executing==2.2.1
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.10.0
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
ipykernel==7.1.0
ipython==9.7.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jieba==0.42.1
jinja2==3.1.6
jupyter-client==8.6.3
jupyter-core==5.9.1
kiwisolver==1.4.9
markupsafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mlx==0.30.0
mlx-lm==0.28.3
mlx-metal==0.30.0
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.4
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
polars==1.35.2
polars-runtime-32==1.35.2
prompt-toolkit==3.0.52
protobuf==6.33.1
psutil==7.1.3
ptyprocess==0.7.0
pure-eval==0.2.3
pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.3
pyzmq==27.1.0
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
stack-data==0.6.3
sympy==1.14.0
tokenizers==0.22.1
torch==2.9.1
torchvision==0.24.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
typing-extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.229
ultralytics-thop==2.0.18
urllib3==2.5.0
wcwidth==0.2.14
